requirements.txt
fastapi
uvicorn[standard]
openai
supabase
python-dotenv
streamlit
httpx
pydantic